Move workflow interfaces to model-nterface
